Detect whether the device is reachable locally
When connecting to a device, mchfuse try to determine if it is locally
reachable, otherwise it uses the external endpoint.

The reachability check runs every 30 seconds (to allow transition from
External to Internal endpoint in case the device become reachable
locally) or when an API call returns a connection error (to allow a
quick transition from Internal to External endpoint if the device
become unreachable locally).